requirements.txt
requests>=2.28.1
beautifulsoup4>=4.11.1
colorama>=0.4.5
argparse>=1.4.0
rich>=12.5.1
urllib3>=1.26.12
tqdm>=4.64.0
pyyaml==6.0.2
python-nmap>=0.7.1
dnspython>=2.2.1
aiohttp>=3.8.3
cryptography>=38.0.1
flask==2.3.3 